Updated all "code/" folder with latest version in cluster [18-01-2026]

+++ home/apujols/code/labeling.py
@@ -11,7 +11,7 @@ import pandas as pd
 
 from pathlib import Path
 
-from preprocessing import update_dataset
+from image_preprocessing import update_dataset
 
 
 def get_filenames_to_label(input_folder):
@@ -19,7 +19,8 @@ def get_filenames_to_label(input_folder):
     zip_files = glob.glob(f"{input_folder}/*.7z")
     for zf in zip_files:
         with py7zr.SevenZipFile(zf, mode='r') as archive:
-            archive.extractall(path=input_folder)
+            path = Path(input_folder).resolve()
+            archive.extractall(path=path)
         os.remove(zf)
 
     # Collect only the .avi and .xml files, delete the rest
@@ -30,7 +31,7 @@ def get_filenames_to_label(input_folder):
 
 def label_as(dataframe, filenames_list, filepath_list, label):
 
-    print(filenames_list)
+    print(f"Files to label: {len(filenames_list)}")
     files_to_process = []
 
     for (file, path) in zip(filenames_list, filepath_list):        
@@ -39,7 +40,7 @@ def label_as(dataframe, filenames_list, filepath_list, label):
             print(f"⚠️ file {file} is not in the dataframe!")
 
             # Process it if possible
-            files_to_process.append(file)
+            files_to_process.append(Path(path))
 
         elif rows.shape[0] > 1:
             print(f"⚠️ The file {file} appears twice in the dataframe.")
@@ -48,12 +49,21 @@ def label_as(dataframe, filenames_list, filepath_list, label):
 
     return dataframe, files_to_process
 
-def clear_incoming_folder(path_incoming_folder):
-    for f in Path(path_incoming_folder).iterdir():
+def clear_incoming_folder(path_incoming_folder, files_to_leave):
+    
+    # Get the stems of the files to leave
+    files_to_leave_names = [f.stem for f in files_to_leave]
+
+    for f in Path(path_incoming_folder).rglob("*"):  # recursively go through all files/folders
         if f.is_file():
-            f.unlink()          # delete file
+            if f.stem not in files_to_leave_names:
+                f.unlink()  # delete the file
         elif f.is_dir():
-            shutil.rmtree(f)    # delete folder recursively
+            # delete empty directories
+            try:
+                f.rmdir()  # only deletes if empty
+            except OSError:
+                pass  # directory not empty, leave it    # delete folder recursively
 
 
 if __name__ == "__main__":
@@ -75,25 +85,24 @@ if __name__ == "__main__":
         filenames, filepaths = get_filenames_to_label(input_folder=incoming_folder)
         dataset, files_to_process = label_as(dataframe=dataset, filenames_list=filenames, filepath_list=filepaths, label=LABEL)
 
+        print(f"Files to process: {len(files_to_process)}")
+        clear_incoming_folder(path_incoming_folder=incoming_folder, files_to_leave=files_to_process)
+
         # Process missing files
         dataset = update_dataset(dataset=dataset,
                                  input_folders=[incoming_folder, incoming_folder],
                                  output_folder=output_folder, 
                                  are_meteors=(True if LABEL == "meteor" else False))
 
+        clear_incoming_folder(path_incoming_folder=incoming_folder, files_to_leave=[])
+        
+        dataset = dataset.drop_duplicates(subset = ['filename'], keep='last')
+        print(f"unique={len(list(dataset['filename'].unique()))}")
+        print(f"num_rows={dataset.shape[0]}")
+
         dataset.to_csv(csv_data_path, sep=";", index=False)
 
-        clear_incoming_folder(path_incoming_folder=incoming_folder)
     else:
         print("⚠️ CSV doesn't exist!")
-
-    # if os.path.exists(csv_data_path):
-    #     temp_df = pd.read_csv(csv_data_path, sep=";")
-    # else:
-    #     temp_df = create_dataset(raw_data_folder=raw_data_folder, output_folder=output_folder, are_meteors=True)
-
-    # df = update_dataset(dataset=temp_df, input_folders=[incoming_folder, incoming_folder], output_folder=output_folder, are_meteors=True)
-
-    # df.to_csv(csv_data_path, sep=";", index=False)
 
     print("DONE!")
